Share inline only pipeline used by InclusionExtension
