+++ src/Microsoft.DocAsCode.MarkdigEngine.Extensions/Inclusion/InclusionExtension.cs
@@ -3,6 +3,7 @@
 
 namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
 {
+    using System.Threading;
     using Markdig;
     using Markdig.Parsers;
     using Markdig.Parsers.Inlines;
@@ -16,6 +17,7 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
     public class InclusionExtension : IMarkdownExtension
     {
         private readonly MarkdownContext _context;
+        private MarkdownPipeline _inlinePipeline;
 
         public InclusionExtension(MarkdownContext context)
         {
@@ -36,7 +38,9 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
             {
                 if (!htmlRenderer.ObjectRenderers.Contains<HtmlInclusionInlineRenderer>())
                 {
-                    htmlRenderer.ObjectRenderers.Insert(0, new HtmlInclusionInlineRenderer(_context, pipeline));
+                    var inlinePipeline = LazyInitializer.EnsureInitialized(ref _inlinePipeline, () => CreateInlineOnlyPipeline(pipeline));
+
+                    htmlRenderer.ObjectRenderers.Insert(0, new HtmlInclusionInlineRenderer(_context, inlinePipeline));
                 }
 
                 if (!htmlRenderer.ObjectRenderers.Contains<HtmlInclusionBlockRenderer>())
@@ -84,6 +88,20 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
                     linkInline.GetDynamicUrl = () => context.GetLink(linkInline.Url, InclusionContext.File, InclusionContext.RootFile);
                 }
             }
+        }
+
+        private static MarkdownPipeline CreateInlineOnlyPipeline(MarkdownPipeline pipeline)
+        {
+            var builder = new MarkdownPipelineBuilder();
+
+            foreach (var extension in pipeline.Extensions)
+            {
+                builder.Extensions.Add(extension);
+            }
+
+            builder.UseInlineOnly();
+
+            return builder.Build();
         }
     }
 }
